+++ aether-admin/src/main/java/com/xiaobai1226/aether/admin/service/intf/FileService.java
@@ -42,5 +42,5 @@ public interface FileService {
     /**
      * 更新File缩略图
      */
-    Boolean updateFileThumbnail(Integer id, String thumbnail);
+    Boolean updateFileThumbnail(Long id, String thumbnail);
 }
