+++ aether-core/src/main/java/com/xiaobai1226/aether/core/service/impl/WebDavServiceImpl.java
@@ -23,7 +23,7 @@ public class WebDavServiceImpl implements WebDavService {
     private UserService userService;
 
     @Override
-    public Integer checkUsernameAndPassword(String username, String password) {
+    public Long checkUsernameAndPassword(String username, String password) {
         // 根据用户名获取用户信息
         var userDO = userService.getUserByUsername(username);
